// src/app/core/enums/faq-type.enum.ts
export enum FaqTypeEnum {
  EQUIPAMENTOS = 'EQUIPAMENTOS',
  INSPECAO = 'INSPECAO',
  CALDEIRA = 'CALDEIRA',
  TESTES = 'TESTES',
  PROFISSIONAL = 'PROFISSIONAL',
  SEGURANCA = 'SEGURANCA',
  SOLDAGEM = 'SOLDAGEM',
  DOCUMENTACAO = 'DOCUMENTACAO',

  O_QUE_E_LAUDO_NR20 = 'O_QUE_E_LAUDO_NR20',
  QUEM_PRECISA_PRONTUARIO = 'QUEM_PRECISA_PRONTUARIO',
  CLASSES_NR20 = 'CLASSES_NR20',
  NIVEIS_TREINAMENTO = 'NIVEIS_TREINAMENTO',
  DIFERENCA_AVANCADO = 'DIFERENCA_AVANCADO',
  FREQUENCIA_ANALISE_RISCO = 'FREQUENCIA_ANALISE_RISCO',
  PERMISSAO_TRABALHO = 'PERMISSAO_TRABALHO',
  DIFERENCA_INFLAMAVEL = 'DIFERENCA_INFLAMAVEL',
  SEM_PRONTUARIO = 'SEM_PRONTUARIO',
  GAMBETI_PRONTUARIO = 'GAMBETI_PRONTUARIO',
  POSTOS_NR20 = 'POSTOS_NR20'


}